ChemReact/requirements.txt
Pillow>=10.0.0
rdkit-pypi>=2022.9.5

# Atom mapping
rxnmapper>=0.4.0
